Moved `useAnnotationQuoteIdle` hook to the folder
