+++ packages/text-annotator-react/src/TextAnnotationPopup/TextAnnotationPopup.tsx
@@ -23,7 +23,7 @@ import {
   useRole
 } from '@floating-ui/react';
 import { isMobile } from './isMobile';
-import { useAnnotationQuoteIdle } from './useAnnotationQuoteIdle';
+import { useAnnotationQuoteIdle } from '../hooks';
 
 import './TextAnnotationPopup.css';
