+++ packages/text-annotator-react/src/TextAnnotationPopup/index.ts
@@ -1,2 +1,1 @@
 export * from './TextAnnotationPopup';
-export * from './useAnnotationQuoteIdle';
