+++ packages/text-annotator-react/src/hooks/useAnnotationQuoteIdle.ts
